move to mineskin package
