+++ src/main/java/de/oliver/fancynpcs/skins/SkinManagerImpl.java
@@ -8,6 +8,7 @@ import de.oliver.fancynpcs.api.skins.SkinGeneratedEvent;
 import de.oliver.fancynpcs.api.skins.SkinManager;
 import de.oliver.fancynpcs.skins.cache.SkinCache;
 import de.oliver.fancynpcs.skins.cache.SkinCacheData;
+import de.oliver.fancynpcs.skins.mineskin.MineSkinQueue;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.lushplugins.chatcolorhandler.ChatColorHandler;
